move methods to Oso
