+++ languages/js/src/Oso.ts
@@ -1,7 +1,10 @@
 import { Polar } from './Polar';
 import { Variable } from './Variable';
-import type { Options, CustomError } from './types';
+import { Expression } from './Expression';
+import { Pattern } from './Pattern';
+import type { Options, CustomError, obj } from './types';
 import { NotFoundError, ForbiddenError, OsoError } from './errors';
+import { serializeTypes, filterData } from './dataFiltering';
 
 /** The Oso authorization API. */
 export class Oso<
@@ -229,5 +232,57 @@ export class Oso<
       fields.add(field);
     }
     return fields;
+  }
+
+  /**
+   * Returns all the resources the actor is allowed to perform some action on.
+   */
+  async authorizedQuery(actor: any, action: any, cls: any): Promise<any> {
+    const resource = new Variable('resource');
+    const host = this.getHost();
+    const clsName = host.types.get(cls)!.name;
+    const constraint = new Expression('And', [
+      new Expression('Isa', [
+        resource,
+        new Pattern({ tag: clsName, fields: {} }),
+      ]),
+    ]);
+    let bindings = new Map();
+    bindings.set('resource', constraint);
+    let results = this.queryRuleWithBindings(
+      'allow',
+      bindings,
+      actor,
+      action,
+      resource
+    );
+
+    const queryResults = [];
+    for await (const result of results) {
+      queryResults.push(result);
+    }
+
+    let jsonResults = queryResults.map(result => ({
+      // `Map<string, any> -> {[key: string]: PolarTerm}` b/c Maps aren't
+      // trivially `JSON.stringify()`-able.
+      bindings: [...result.entries()].reduce((obj: obj, [k, v]) => {
+        obj[k] = host.toPolar(v);
+        return obj;
+      }, {}),
+    }));
+    let resultsStr = JSON.stringify(jsonResults);
+    let typesStr = serializeTypes(host.types);
+    let plan = this.getFfi().buildFilterPlan(
+      typesStr,
+      resultsStr,
+      'resource',
+      clsName
+    );
+    return filterData(host, plan);
+  }
+
+  async authorizedResources(actr: any, actn: any, cls: any): Promise<any> {
+    const query = await this.authorizedQuery(actr, actn, cls);
+    return !query ? [] : this.getHost().types.get(cls)!.execQuery!(query);
   }
 }
